LayoutTests/platform/glib/imported/w3c/web-platform-tests/cookie-store/change_eventhandler_for_no_name_equals_in_value.https.window-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: TypeError: The cookie name and value must not both be empty.

FAIL Verify that attempting to set a cookie with no name and with '=' in the value does not work. promise_test: Unhandled rejection with value: object "TypeError: The cookie name and value must not both be empty."
